Zurmo 0.6.6

+++ app/protected/modules/import/models/Import.php
@@ -108,9 +108,13 @@
 
         protected function beforeDelete()
         {
-            parent::beforeDelete();
+            if (!parent::beforeDelete())
+            {
+                return false;
+            }
             $sql = 'Drop table if exists ' . $this->getTempTableName();
             R::exec($sql);
+            return true;
         }
     }
 ?>
